Property to disable deleting old comments from pull requests

+++ src/main/java/org/sonar/plugins/github/PullRequestIssuePostJob.java
@@ -66,7 +66,9 @@ public class PullRequestIssuePostJob implements PostJob {
 
       updateReviewComments(commentsToBeAddedByLine);
 
-      pullRequestFacade.deleteOutdatedComments();
+      if (gitHubPluginConfiguration.isDeleteOldCommentsEnabled()) {
+        pullRequestFacade.deleteOutdatedComments();
+      }
 
       pullRequestFacade.createOrUpdateGlobalComments(report.hasNewIssue() ? report.formatForMarkdown() : null);
